move wasm module packaging tests into the separate file
