+++ hermes/bin/src/packaging/module/mod.rs
@@ -4,6 +4,8 @@ mod author_payload;
 mod config;
 mod manifest;
 mod settings;
+#[cfg(test)]
+pub(crate) mod tests;
 
 pub(crate) use author_payload::{SignaturePayload, SignaturePayloadBuilder};
 use chrono::{DateTime, Utc};
@@ -415,234 +417,4 @@ pub(crate) fn write_share_dir(
     let share_dir = dir.create_dir(path)?;
     share_dir.copy_resource_dir(resource, &Path::default())?;
     Ok(())
-}
-
-#[cfg(test)]
-pub(crate) mod tests {
-    use temp_dir::TempDir;
-
-    use super::*;
-    use crate::{
-        hdf5::resources::{FsResource, ResourceBuilder},
-        packaging::sign::{
-            certificate::{self, tests::certificate_str},
-            keys::tests::private_key_str,
-        },
-    };
-
-    pub(crate) struct ModulePackageFiles {
-        pub(crate) metadata: Metadata<ModulePackage>,
-        pub(crate) component: Vec<u8>,
-        pub(crate) config_schema: ConfigSchema,
-        pub(crate) config: Config,
-        pub(crate) settings_schema: SettingsSchema,
-    }
-
-    #[allow(clippy::unwrap_used)]
-    pub(crate) fn prepare_default_package_files() -> ModulePackageFiles {
-        let metadata = Metadata::<ModulePackage>::from_reader(
-            serde_json::json!(
-                {
-                    "$schema": "https://raw.githubusercontent.com/input-output-hk/hermes/main/hermes/schemas/hermes_module_metadata.schema.json",
-                    "name": "Test module",
-                    "version": "V1.0.0",
-                    "description": "Some description",
-                    "src": ["https://github.com/input-output-hk/hermes"],
-                    "copyright": ["Copyright Ⓒ 2024, IOG Singapore."],
-                    "license": [{"spdx": "MIT"}]
-                }
-            ).to_string().as_bytes(),
-        ).unwrap();
-        let config_schema =
-            ConfigSchema::from_reader(serde_json::json!({}).to_string().as_bytes()).unwrap();
-
-        let config = Config::from_reader(
-            serde_json::json!({}).to_string().as_bytes(),
-            config_schema.validator(),
-        )
-        .unwrap();
-
-        let settings_schema =
-            SettingsSchema::from_reader(serde_json::json!({}).to_string().as_bytes()).unwrap();
-
-        let component = r#"
-            (component
-                (core module $Module
-                    (export "foo" (func $foo))
-                    (func $foo (result i32)
-                        i32.const 1
-                    )
-                )
-                (core instance $module (instantiate (module $Module)))
-                (func $foo (result s32) (canon lift (core func $module "foo")))
-                (export "foo" (func $foo))
-            )"#;
-
-        ModulePackageFiles {
-            metadata,
-            component: component.as_bytes().to_vec(),
-            config_schema,
-            config,
-            settings_schema,
-        }
-    }
-
-    #[allow(clippy::unwrap_used)]
-    pub(crate) fn prepare_package_dir(
-        module_name: String, dir: &TempDir, module_package_files: &ModulePackageFiles,
-    ) -> Manifest {
-        let config_path = dir.path().join("config.json");
-        let config_schema_path = dir.path().join("config.schema.json");
-        let metadata_path = dir.path().join("metadata.json");
-        let component_path = dir.path().join("module.wasm");
-        let settings_schema_path = dir.path().join("settings.schema.json");
-
-        std::fs::write(
-            &metadata_path,
-            module_package_files.metadata.to_bytes().unwrap().as_slice(),
-        )
-        .unwrap();
-        std::fs::write(&component_path, module_package_files.component.as_slice()).unwrap();
-        std::fs::write(
-            &config_path,
-            module_package_files.config.to_bytes().unwrap().as_slice(),
-        )
-        .unwrap();
-        std::fs::write(
-            &config_schema_path,
-            module_package_files
-                .config_schema
-                .to_bytes()
-                .unwrap()
-                .as_slice(),
-        )
-        .unwrap();
-        std::fs::write(
-            &settings_schema_path,
-            module_package_files
-                .settings_schema
-                .to_bytes()
-                .unwrap()
-                .as_slice(),
-        )
-        .unwrap();
-
-        Manifest {
-            name: module_name,
-            metadata: ResourceBuilder::Fs(FsResource::new(metadata_path)),
-            component: ResourceBuilder::Fs(FsResource::new(component_path)),
-            config: manifest::ManifestConfig {
-                file: Some(ResourceBuilder::Fs(FsResource::new(config_path))),
-                schema: ResourceBuilder::Fs(FsResource::new(config_schema_path)),
-            }
-            .into(),
-            settings: manifest::ManifestSettings {
-                schema: ResourceBuilder::Fs(FsResource::new(settings_schema_path)),
-            }
-            .into(),
-            share: None,
-        }
-    }
-
-    #[allow(clippy::unwrap_used)]
-    pub(crate) fn check_module_integrity(
-        module_files: &ModulePackageFiles, module_package: &ModulePackage,
-    ) {
-        let package_metadata = module_package.get_metadata().unwrap();
-        assert_eq!(module_files.metadata, package_metadata);
-
-        // check WASM component file
-        assert!(module_package.get_component().is_ok());
-
-        // check config and config schema JSON files
-        let (package_config, package_config_schema) =
-            module_package.get_config_with_schema().unwrap();
-        assert_eq!(module_files.config, package_config.unwrap());
-        assert_eq!(module_files.config_schema, package_config_schema.unwrap());
-
-        // check settings schema JSON file
-        let package_settings_schema = module_package.get_settings_schema().unwrap();
-        assert_eq!(
-            module_files.settings_schema,
-            package_settings_schema.unwrap()
-        );
-    }
-
-    #[test]
-    #[allow(clippy::unwrap_used)]
-    fn from_dir_test() {
-        let dir = TempDir::new().unwrap();
-
-        let mut module_package_files = prepare_default_package_files();
-
-        let manifest = prepare_package_dir("module".to_string(), &dir, &module_package_files);
-
-        let build_time = DateTime::default();
-        let package =
-            ModulePackage::build_from_manifest(&manifest, dir.path(), None, build_time).unwrap();
-
-        assert!(package.validate(true).is_ok());
-
-        // Module package during the build process updates metadata file
-        // to have a corresponded values update `module_package_files`.
-        module_package_files.metadata.set_name(&manifest.name);
-        module_package_files.metadata.set_build_date(build_time);
-
-        check_module_integrity(&module_package_files, &package);
-    }
-
-    #[test]
-    #[allow(clippy::unwrap_used)]
-    fn sign_test() {
-        let dir = TempDir::new().unwrap();
-
-        let mut module_package_files = prepare_default_package_files();
-
-        let manifest = prepare_package_dir("module".to_string(), &dir, &module_package_files);
-
-        let build_time = DateTime::default();
-        let package =
-            ModulePackage::build_from_manifest(&manifest, dir.path(), None, build_time).unwrap();
-
-        assert!(package.validate(true).is_ok());
-        assert!(package.validate(false).is_err());
-        assert!(package.get_signature().unwrap().is_none());
-
-        let private_key = PrivateKey::from_str(&private_key_str()).unwrap();
-        let certificate = Certificate::from_str(&certificate_str()).unwrap();
-        package.sign(&private_key, &certificate).unwrap();
-        package.sign(&private_key, &certificate).unwrap();
-
-        assert!(package.get_signature().unwrap().is_some());
-
-        assert!(
-            package.validate(false).is_err(),
-            "Missing certificate in the storage."
-        );
-
-        certificate::storage::add_certificate(certificate).unwrap();
-        assert!(package.validate(false).is_ok());
-
-        // corrupt payload with the modifying metadata.json file
-        module_package_files.metadata.set_name("New name");
-        package
-            .0
-            .remove_file(ModulePackage::METADATA_FILE.into())
-            .unwrap();
-        package
-            .0
-            .copy_resource_file(
-                &BytesResource::new(
-                    ModulePackage::METADATA_FILE.to_string(),
-                    module_package_files.metadata.to_bytes().unwrap(),
-                ),
-                ModulePackage::METADATA_FILE.into(),
-            )
-            .unwrap();
-
-        assert!(
-            package.validate(false).is_err(),
-            "Corrupted signature payload."
-        );
-    }
 }
